// src/app/models/generated.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string
    String: string
    Boolean: boolean
    Int: number
    Float: number
    /** The `Upload` scalar type represents a file upload. */
    Upload: any
}

export enum CacheControlScope {
    Private = 'PRIVATE',
    Public = 'PUBLIC'
}

export type Character = {
    __typename?: 'Character'
    /** Time at which the character was created in the database. */
    created?: Maybe<Scalars['String']>
    /** Episodes in which this character appeared. */
    episode: Array<Maybe<Episode>>
    /** The gender of the character ('Female', 'Male', 'Genderless' or 'unknown'). */
    gender?: Maybe<Scalars['String']>
    /** The id of the character. */
    id?: Maybe<Scalars['ID']>
    /**
     * Link to the character's image.
     * All images are 300x300px and most are medium shots or portraits since they are intended to be used as avatars.
     */
    image?: Maybe<Scalars['String']>
    /** The character's last known location */
    location?: Maybe<Location>
    /** The name of the character. */
    name?: Maybe<Scalars['String']>
    /** The character's origin location */
    origin?: Maybe<Location>
    /** The species of the character. */
    species?: Maybe<Scalars['String']>
    /** The status of the character ('Alive', 'Dead' or 'unknown'). */
    status?: Maybe<Scalars['String']>
    /** The type or subspecies of the character. */
    type?: Maybe<Scalars['String']>
}

export type Characters = {
    __typename?: 'Characters'
    info?: Maybe<Info>
    results?: Maybe<Array<Maybe<Character>>>
}

export type Episode = {
    __typename?: 'Episode'
    /** The air date of the episode. */
    air_date?: Maybe<Scalars['String']>
    /** List of characters who have been seen in the episode. */
    characters: Array<Maybe<Character>>
    /** Time at which the episode was created in the database. */
    created?: Maybe<Scalars['String']>
    /** The code of the episode. */
    episode?: Maybe<Scalars['String']>
    /** The id of the episode. */
    id?: Maybe<Scalars['ID']>
    /** The name of the episode. */
    name?: Maybe<Scalars['String']>
}
// ...
export type Info = {
    __typename?: 'Info'
    /** The length of the response. */
    count?: Maybe<Scalars['Int']>
    /** Number of the next page (if it exists) */
    next?: Maybe<Scalars['Int']>
    /** The amount of pages. */
    pages?: Maybe<Scalars['Int']>
    /** Number of the previous page (if it exists) */
    prev?: Maybe<Scalars['Int']>
}

export type Location = {
    __typename?: 'Location'
    /** Time at which the location was created in the database. */
    created?: Maybe<Scalars['String']>
    /** The dimension in which the location is located. */
    dimension?: Maybe<Scalars['String']>
    /** The id of the location. */
    id?: Maybe<Scalars['ID']>
    /** The name of the location. */
    name?: Maybe<Scalars['String']>
    /** List of characters who have been last seen in the location. */
    residents: Array<Maybe<Character>>
    /** The type of the location. */
    type?: Maybe<Scalars['String']>
}